Move duplicateQualityChanges() to QualityManager
